screen: Use better approach for toggling fullscreen layout
Bind F1 to 'exec' command to execute a shellscript that has a bunch of
screen -X commands in a subprocess. When closing fullscreen with
Ctrl+F1, user still have to manually type in the last layout which
should be shown on message line.
Caveat: If there are other screen messages show up during fullscreen,
the last layout message would not be shown correctly.